refacter: 페이지 정리 및 컴포넌트 구조화
