+++ src/Pages/EventVotePage.jsx
@@ -1,51 +1,16 @@
 import { useParams } from 'react-router-dom'
-import { useRef, useState } from 'react'
+import { useState } from 'react'
 import useVote from '../api/useVote.js'
-import PodiumModal from '../components/PodiumModal.jsx'
 import '../style/VotePage.css'
-import truckImg from '../data/icon/truck.png'
-import useOnClickOutside from '../hooks/useOnClickOutside.js'
 import { useEvent } from './EventPage.jsx'
 
-function classifyMenu(name) {
-  const lower = name.toLowerCase()
-  if (
-    ['치킨', '닭', '덮밥', '갈비', '불고기', '비빔밥'].some((x) =>
-      lower.includes(x),
-    )
-  )
-    return '한식'
-  if (
-    ['버거', '핫도그', '피자', '파스타', '감자튀김'].some((x) =>
-      lower.includes(x),
-    )
-  )
-    return '양식'
-  if (
-    ['라멘', '우동', '오코노미야끼', '타코야끼'].some((x) => lower.includes(x))
-  )
-    return '일식'
-  if (['타코', '나쵸', '케사디야', '케밥'].some((x) => lower.includes(x)))
-    return '멕시칸'
-  if (['붕어빵', '어묵', '떡볶이', '순대'].some((x) => lower.includes(x)))
-    return '분식'
-  if (
-    ['티라미수', '브라우니', '라떼', '커피', '딸기'].some((x) =>
-      lower.includes(x),
-    )
-  )
-    return '디저트'
-  return '기타'
-}
-
-function classifyTruckCategory(truck) {
-  const menuNames = truck.menus?.map((m) => m.menuName) || []
-  for (const name of menuNames) {
-    const category = classifyMenu(name)
-    if (category !== '기타') return category
-  }
-  return '기타'
-}
+import {
+  RankedTruck,
+  MenuDetail,
+  TruckCard,
+  TruckNotFound,
+  VoteAfterModal,
+} from '../components/index.jsx'
 
 export default function EventVotePage() {
   const { eventId } = useParams()
@@ -56,7 +21,6 @@ export default function EventVotePage() {
     setVotedTruckIds,
     fetchVoteResult,
   } = useEvent()
-  const [showPodium, setShowPodium] = useState(false)
   const { vote } = useVote(eventId)
 
   const [activeCategory, setActiveCategory] = useState('전체')
@@ -106,9 +70,7 @@ export default function EventVotePage() {
   const filteredTrucks =
     activeCategory === '전체'
       ? merged || []
-      : merged?.filter(
-          (truck) => classifyTruckCategory(truck) === activeCategory,
-        )
+      : merged?.filter((truck) => truck?.category === activeCategory)
 
   return (
     <div>
@@ -123,12 +85,6 @@ export default function EventVotePage() {
         )}
         <hr className="event-divider" />
       </div>
-      {/* vote animation */}
-      {/* <VoteRanking
-        setShowPodium={setShowPodium}
-        showPodium={showPodium}
-        sorted={sorted}
-      /> */}
 
       <div className="text-center mb-16">
         <h2 className="text-4xl font-extrabold mb-4 text-indigo-800 tracking-tight inline-block bg-clip-text text-transparent bg-gradient-to-r from-indigo-600 to-purple-600">
@@ -180,443 +136,6 @@ export default function EventVotePage() {
       )}
       {/* 투표 성공 모달 */}
       {showModal && <VoteAfterModal />}
-    </div>
-  )
-}
-
-const TruckNotFound = () => {
-  return (
-    <div className="h-[500px] flex justify-center items-center">
-      <div className="relative w-[300px] h-24 flex flex-col gap-3">
-        <img
-          src={truckImg}
-          className="animate-[truckMove_2s_linear_forwards] w-[100px]"
-        />
-        <h2 className="text-2xl">트럭을 찾을 수 없습니다</h2>
-      </div>
-    </div>
-  )
-}
-
-const TruckCard = ({ truck, handleVote, isVoted }) => {
-  const [expandedMenus, setExpandedMenus] = useState([])
-  const toggleMenu = (truckId) => {
-    if (expandedMenus.includes(truckId)) {
-      setExpandedMenus(expandedMenus.filter((id) => id !== truckId))
-    } else {
-      setExpandedMenus([...expandedMenus, truckId])
-    }
-  }
-  return (
-    <div
-      key={truck.truckId}
-      className="bg-white rounded-2xl overflow-hidden 
-  shadow-lg hover:shadow-xl transition-all duration-300 
-  transform hover:-translate-y-2 border border-indigo-50 flex flex-col"
-    >
-      <div className="h-56 overflow-hidden relative group">
-        <img
-          src={truck?.menus[1]?.menuImage}
-          alt={truck?.truckName + ' 이미지'}
-          className="w-full h-full object-cover object-top transition-transform duration-700 group-hover:scale-110"
-        />
-        <div className="absolute inset-0 bg-gradient-to-t from-indigo-900/70 to-transparent opacity-0 group-hover:opacity-100 transition-opacity duration-300 flex items-end">
-          <div className="p-4 text-white">
-            <p className="font-medium">{truck?.description}</p>
-          </div>
-        </div>
-        <div className="absolute top-4 right-4">
-          <span className="bg-white/90 backdrop-blur-sm text-indigo-800 text-xs font-bold px-3 py-1.5 rounded-full shadow-md">
-            {'한식'}
-          </span>
-        </div>
-      </div>
-
-      <div className="p-6">
-        <h3 className="text-2xl font-bold text-gray-800 mb-3">
-          {truck.truckName}
-        </h3>
-
-        <div className="mb-6">
-          <button
-            onClick={() => toggleMenu(truck.truckId)}
-            className="w-full bg-gradient-to-r from-indigo-50 to-purple-50 hover:from-indigo-100 hover:to-purple-100 px-4 py-3 rounded-xl text-left transition-all duration-300 flex justify-between items-center"
-          >
-            <span className="font-semibold text-gray-700 flex items-center">
-              <i className="fas fa-utensils text-indigo-500 mr-2"></i>
-              메뉴 보기
-            </span>
-            <i
-              className={`fas fa-chevron-down text-indigo-500 transition-transform duration-300 ${
-                expandedMenus.includes(truck.truckId) ? 'rotate-180' : ''
-              }`}
-            ></i>
-          </button>
-
-          <div
-            className={`overflow-hidden transition-all duration-300 ${
-              expandedMenus.includes(truck.truckId)
-                ? 'max-h-[500px] mt-4'
-                : 'max-h-0'
-            }`}
-          >
-            <div className="grid grid-cols-1 gap-4">
-              {truck.menus?.map((item, idx) => (
-                <div
-                  key={idx}
-                  className="bg-white p-3 rounded-lg shadow-sm border border-indigo-50 
-              hover:border-indigo-200 transition-colors"
-                >
-                  <div className="flex items-center">
-                    <div className="w-16 h-16 bg-indigo-100 rounded-lg overflow-hidden mr-4">
-                      <img
-                        src={item.menuImage}
-                        alt={'메뉴 이미지'}
-                        className="w-full h-full object-cover"
-                      />
-                    </div>
-                    <div className="flex-1">
-                      <h4 className="font-bold text-gray-800">
-                        {item.menuName}
-                      </h4>
-                      <p className="text-indigo-600 font-medium">
-                        {item.menuPrice}
-                      </p>
-                      <p className="text-gray-500 text-sm line-clamp-1">
-                        {item.description}
-                      </p>
-                    </div>
-                  </div>
-                </div>
-              ))}
-            </div>
-          </div>
-        </div>
-
-        <div className="flex justify-between items-center">
-          <div className="text-indigo-700 bg-indigo-50 px-4 py-2 rounded-full">
-            <span className="font-bold text-lg">{truck.voteCount}</span>
-            <span className="text-sm ml-1">투표</span>
-          </div>
-          <button
-            onClick={() => handleVote(truck.truckId)}
-            className={`px-5 py-2.5 rounded-lg transition-all duration-300 font-medium whitespace-nowrap cursor-pointer ${
-              isVoted
-                ? 'bg-gray-200 text-gray-600'
-                : 'bg-gradient-to-r from-indigo-600 to-purple-600 hover:from-indigo-700 hover:to-purple-700 text-white shadow-md hover:shadow-lg'
-            }`}
-            disabled={isVoted}
-          >
-            {isVoted ? (
-              <>
-                <i className="fas fa-check-circle mr-2"></i>투표 완료
-              </>
-            ) : (
-              <>
-                <i className="fas fa-vote-yea mr-2"></i>투표하기
-              </>
-            )}
-          </button>
-        </div>
-      </div>
-    </div>
-  )
-}
-
-const RankedTruck = ({ showMenuDetail, sorted, selectedTruck }) => {
-  console.log('ranked', sorted)
-  return (
-    <div>
-      <div className="text-center mb-16">
-        <h2 className="text-4xl font-extrabold mb-4 text-indigo-800 tracking-tight inline-block bg-clip-text text-transparent bg-gradient-to-r from-indigo-600 to-purple-600">
-          인기
-        </h2>
-        <div className="w-24 h-1 bg-gradient-to-r from-indigo-600 to-purple-600 mx-auto mb-6 rounded-full"></div>
-      </div>
-      <div className="grid grid-cols-1 md:grid-cols-3 gap-8">
-        {sorted.slice(0, 3).map((truck, index) => (
-          <RankedTruckCard
-            truck={truck}
-            index={index}
-            key={index}
-            selectedTruck={selectedTruck}
-            showMenuDetail={showMenuDetail}
-          />
-        ))}
-      </div>
-    </div>
-  )
-}
-
-const RankedTruckCard = ({ truck, index, selectedTruck, showMenuDetail }) => {
-  console.log('card', truck)
-  return (
-    <div
-      key={truck?.truckId}
-      className={`bg-white rounded-2xl overflow-hidden shadow-xl border-2 ${
-        index === 0
-          ? 'border-yellow-400  z-10'
-          : index === 1
-            ? 'border-gray-300'
-            : index === 2
-              ? 'border-amber-600'
-              : 'border-blue-700'
-      }`}
-    >
-      <div className="relative">
-        <img
-          src={truck?.menuImage ?? truck?.menus[1]?.menuImage}
-          alt={'트럭 이미지'}
-          className="w-full h-56 object-cover object-top"
-        />
-        <div
-          className={`absolute top-4 left-4 w-10 h-10 rounded-full flex items-center justify-center font-bold text-white shadow-lg ${
-            index === 0
-              ? 'bg-yellow-500'
-              : index === 1
-                ? 'bg-gray-400'
-                : index === 2
-                  ? 'bg-amber-700'
-                  : 'bg-blue-700'
-          }`}
-        >
-          {index + 1}
-        </div>
-
-        {index <= 2 && (
-          <div>
-            <div
-              className={`absolute top-0 right-0 w-0 h-0 border-t-[80px] ${
-                index === 0
-                  ? 'border-t-yellow-500'
-                  : index === 1
-                    ? 'border-t-gray-400'
-                    : 'border-t-amber-700'
-              } border-l-[80px] border-l-transparent`}
-            ></div>
-            <div className="absolute top-3 right-3 text-white font-bold transform rotate-45">
-              TOP
-            </div>
-          </div>
-        )}
-      </div>
-      <div className="p-6">
-        <h4 className="text-xl font-bold text-gray-800 mb-3">
-          {truck?.truckName}
-        </h4>
-        <div className="flex justify-between items-center">
-          <span className="bg-indigo-50 text-indigo-700 px-3 py-1 rounded-full text-sm font-medium">
-            {truck?.category ?? '한식'}
-          </span>
-          <span
-            className={`font-bold text-lg ${
-              index === 0
-                ? 'text-yellow-500'
-                : index === 1
-                  ? 'text-gray-400'
-                  : index === 2
-                    ? 'text-amber-700'
-                    : 'text-blue-700'
-            }`}
-          >
-            {truck?.voteCount} 투표
-          </span>
-        </div>
-
-        {/* 메뉴 미리보기 추가 */}
-        <div className="mt-4 pt-4 border-t border-indigo-50">
-          <div className="flex overflow-x-auto pb-2 space-x-3">
-            {selectedTruck?.menus?.map((item, idx) => (
-              <div key={idx} className="flex-shrink-0 w-20">
-                <div className="w-20 h-20 rounded-lg overflow-hidden mb-1">
-                  <img
-                    src={item.image}
-                    alt={item.name}
-                    className="w-full h-full object-cover"
-                  />
-                </div>
-                <p className="text-xs text-center font-medium text-gray-700 truncate">
-                  {item.name}
-                </p>
-              </div>
-            ))}
-          </div>
-        </div>
-
-        <button
-          onClick={() => {
-            showMenuDetail(truck)
-          }}
-          className="w-full mt-4 py-2 bg-gradient-to-r from-indigo-50
-       to-purple-50 hover:from-indigo-100 hover:to-purple-100 rounded-lg text-indigo-700 font-medium flex items-center justify-center cursor-pointer"
-        >
-          <i className="fas fa-utensils mr-2"></i>
-          메뉴 자세히 보기
-        </button>
-      </div>
-    </div>
-  )
-}
-
-const VoteRanking = ({ showPodium, setShowPodium, sorted }) => {
-  return (
-    <div className="podium-container">
-      <h1 className="porium-vote-title">이벤트 투표 결과</h1>
-      <div className="arrow-group">
-        <div className="arrow">↘ </div>
-        <div className="arrow">↓ </div>
-        <div className="arrow"> ↙</div>
-      </div>
-      <div className="result-btn-container">
-        <button onClick={() => setShowPodium(true)} className="result-btn">
-          현재 1위는 <span className="question-mark">?</span>
-        </button>
-      </div>
-      {showPodium && (
-        <PodiumModal results={sorted} onClose={() => setShowPodium(false)} />
-      )}
-    </div>
-  )
-}
-
-const VoteAfterModal = () => {
-  return (
-    <div className="fixed inset-0 flex items-center justify-center z-50 pointer-events-none">
-      <div className="bg-white/90 backdrop-blur-sm px-8 py-6 rounded-xl shadow-lg border-2 border-indigo-500 transform transition-all duration-300 animate-bounce">
-        <div className="text-center">
-          <div className="w-16 h-16 bg-indigo-100 rounded-full flex items-center justify-center mx-auto mb-4">
-            <i className="fas fa-check-circle text-indigo-600 text-3xl"></i>
-          </div>
-          <h3 className="text-xl font-bold text-indigo-800 mb-2">투표 완료!</h3>
-          <p className="text-gray-600">소중한 의견 감사합니다.</p>
-        </div>
-      </div>
-    </div>
-  )
-}
-
-const MenuDetail = ({
-  selectedTruck,
-  setSelectTruck,
-  votedTruckIds,
-  handleVote,
-}) => {
-  const ref = useRef()
-  useOnClickOutside(ref, () => {
-    setSelectTruck(null)
-  })
-  return (
-    <div className="fixed inset-0 bg-black/70 flex items-center justify-center z-50 p-4">
-      <div
-        ref={ref}
-        className="bg-white rounded-2xl max-w-4xl w-full max-h-[90vh] overflow-y-auto"
-      >
-        <div className="p-6">
-          <div className="flex justify-between items-center mb-6">
-            <h3 className="text-2xl font-bold text-indigo-800">
-              {selectedTruck.truckName}
-            </h3>
-            <button
-              onClick={() => setSelectTruck(null)}
-              className="text-gray-500 hover:text-gray-700 text-2xl 
-              cursor-pointer"
-            >
-              <i className="fas fa-times">X</i>
-            </button>
-          </div>
-
-          <div className="grid grid-cols-1 md:grid-cols-2 gap-8">
-            <div>
-              <img
-                src={selectedTruck.menus[0].menuImage}
-                alt={selectedTruck.name}
-                className="w-full h-64 object-cover object-top rounded-xl"
-              />
-              <div className="mt-4">
-                <p className="text-gray-600 mb-4">
-                  {selectedTruck.description}
-                </p>
-                <div className="flex items-center">
-                  <span className="bg-indigo-100 text-indigo-800 px-3 py-1 rounded-full text-sm font-medium">
-                    {selectedTruck.category ?? '한식'}
-                  </span>
-                  <span className="ml-4 text-indigo-600 font-medium">
-                    <i className="fas fa-thumbs-up mr-1"></i>{' '}
-                    {selectedTruck.voteCount} 투표
-                  </span>
-                </div>
-              </div>
-            </div>
-
-            <div>
-              <h4 className="text-lg font-bold text-gray-800 mb-4 flex items-center">
-                <i className="fas fa-utensils text-indigo-600 mr-2"></i>
-                대표 메뉴
-              </h4>
-              <div className="space-y-6">
-                {selectedTruck.menus.map((menu, idx) => (
-                  <div
-                    key={idx}
-                    className="bg-white p-4 rounded-xl shadow-md border border-indigo-100"
-                  >
-                    <div className="flex flex-col md:flex-row">
-                      <div className="md:w-1/3 mb-4 md:mb-0">
-                        <img
-                          src={menu.menuImage}
-                          alt={'이미지'}
-                          className="w-full h-40 object-cover rounded-lg"
-                        />
-                      </div>
-                      <div className="md:w-2/3 md:pl-6">
-                        <div className="flex justify-between items-start mb-2">
-                          <h5 className="text-xl font-bold text-gray-800">
-                            {menu.menuName}
-                          </h5>
-                          <span className="text-indigo-700 font-bold">
-                            {menu.menuPrice}원
-                          </span>
-                        </div>
-                        <p className="text-gray-600 mb-4">{menu.description}</p>
-                        <div className="flex items-center text-sm text-gray-500">
-                          <i className="fas fa-fire text-red-500 mr-1"></i>
-                          <span>인기 메뉴</span>
-                        </div>
-                      </div>
-                    </div>
-                  </div>
-                ))}
-              </div>
-
-              <div className="mt-8 flex justify-end">
-                <button
-                  onClick={() => {
-                    setSelectTruck(null)
-                    if (!votedTruckIds.includes(selectedTruck.truckId)) {
-                      handleVote(selectedTruck.truckId)
-                    }
-                  }}
-                  className={`px-6 py-3 rounded-lg transition-all duration-300 font-medium whitespace-nowrap !rounded-button cursor-pointer ${
-                    votedTruckIds.includes(selectedTruck.truckId)
-                      ? 'bg-gray-200 text-gray-600'
-                      : 'bg-gradient-to-r from-indigo-600 to-purple-600 hover:from-indigo-700 hover:to-purple-700 text-white shadow-md hover:shadow-lg'
-                  }`}
-                  disabled={votedTruckIds.includes(selectedTruck.truckId)}
-                >
-                  {votedTruckIds.includes(selectedTruck.truckId) ? (
-                    <>
-                      <i className="fas fa-check-circle mr-2"></i>투표 완료
-                    </>
-                  ) : (
-                    <>
-                      <i className="fas fa-vote-yea mr-2"></i>투표하기
-                    </>
-                  )}
-                </button>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
     </div>
   )
 }
