+++ src/components/index.jsx
@@ -1,0 +1,5 @@
+export { default as MenuDetail } from './Vote/MenuDetail'
+export { default as VoteAfterModal } from './Vote/Modal'
+export { default as TruckNotFound } from './Vote/NotFound'
+export { default as RankedTruck } from './Vote/RankedTruck'
+export { default as TruckCard } from './Vote/TruckCard'
